Move data exception to own package
