+++ src/seedu/addressbook/model/person/Phone.java
@@ -1,6 +1,6 @@
 package seedu.addressbook.model.person;
 
-import seedu.addressbook.model.IllegalValueException;
+import seedu.addressbook.model.exception.IllegalValueException;
 
 /**
  * Represents a Person's phone number in the address book.
