+++ src/seedu/addressbook/parser/Parser.java
@@ -1,7 +1,7 @@
 package seedu.addressbook.parser;
 
 import seedu.addressbook.commands.*;
-import seedu.addressbook.model.IllegalValueException;
+import seedu.addressbook.model.exception.IllegalValueException;
 import seedu.addressbook.model.tag.Tag;
 import seedu.addressbook.model.tag.UniqueTagList;
 import seedu.addressbook.model.person.*;
